requirements.txt
# NewsPocket Dependencies
# RSS Parsing
feedparser>=6.0.10

# HTTP Requests
requests>=2.31.0

# Template Engine
Jinja2>=3.1.2

# Date/Time Handling
python-dateutil>=2.8.2

# Environment Variables (optional, for local dev)
python-dotenv>=1.0.0

# Web Dashboard
streamlit>=1.30.0
plotly>=5.18.0
